+++ portfolioAnalytics/settings.py
@@ -12,7 +12,7 @@
 # either express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Integration Settings."""
+"""Integration Settings for Threshold Based Methods."""
 
 GRID_POINTS = 3000
 PRECISION = 1.e-8