+++ portfolioAnalytics/vasicek.py
@@ -1,7 +1,4 @@
 # encoding: utf-8
-
-"""Implementing the large pool portfolio approximation of Vasicek."""
-
 # (c) 2014-2019 Open Risk (https://www.openriskmanagement.com)
 #
 # portfolioAnalytics is licensed under the Apache 2.0 license a copy of which is included
@@ -14,6 +11,20 @@
 # either express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
+""" Implementing the credit portfolio model of Vasicek.
+
+See `Vasicek Distribution <https://www.openriskmanual.org/wiki/Vasicek_Distribution>`_
+
+* vasicek_base_ implements a finite homogeneous pool
+* vasicek_base_el_ implements the expected loss for the vasicek_base case
+* vasicek_base_ul_ implements the standard deviation for the vasicek_base case
+* vasicek_lim_ implements the limiting case for large N
+* vasicek_lim_el_ implements the expected loss for the vasicek_lim case
+* vasicek_lim_ul_ implements the standard deviation for the vasicek_lim case
+* vasicek_lim_q_ implements the quantile for the vasicek_lim case
+
+"""
+
 import math
 
 from scipy import stats
@@ -25,11 +36,11 @@ from portfolioAnalytics import settings
 def vasicek_base(N, k, p, rho):
     """Vasicek Base Discrete distribution.
 
-    :param N:
-    :param k:
-    :param p:
-    :param rho:
-    :return:
+    :param N: The number of entities in the portfolio
+    :param k: The number of defaults
+    :param p:   The probability of default (uniform across the portfolio)
+    :param rho: The asset correlation parameter
+    :return: The probability of k defaults
     """
 
     zmin = - settings.SCALE
@@ -52,10 +63,10 @@ def vasicek_base(N, k, p, rho):
 def vasicek_base_el(N, p, rho):
     """Expected Loss for the Vasicek Base distribution.
 
-    :param N:
-    :param p:
-    :param rho:
-    :return:
+    :param N: The number of entities in the portfolio
+    :param p:   The probability of default
+    :param rho: The asset correlation (not needed here)
+    :return: The average default rate / loss
     """
     return N * p
 
@@ -63,10 +74,10 @@ def vasicek_base_el(N, p, rho):
 def vasicek_base_ul(N, p, rho):
     """Unexpected Loss for the Vasicek Base distribution.
 
-    :param N:
-    :param p:
-    :param rho:
-    :return:
+    :param N: The number of entities in the portfolio
+    :param p: The probability of default
+    :param rho: The asset correlation (not needed here)
+    :return: The default rate volatility (UL)
     """
     zmin = - settings.SCALE
     zmax = settings.SCALE
@@ -89,10 +100,10 @@ def vasicek_base_ul(N, p, rho):
 def vasicek_lim(theta, p, rho):
     """The Large-N limit of the Vasicek Distribution.
 
-    :param theta:
-    :param p:
-    :param rho:
-    :return:
+    :param theta: The target default rate
+    :param p:   The probability of default
+    :param rho: The asset correlation
+    :return: Cumulative probability
     """
     a1 = stats.norm.ppf(p, loc=0.0, scale=1.0)
     arg1 = stats.norm.ppf(theta, loc=0.0, scale=1.0)
@@ -104,9 +115,9 @@ def vasicek_lim(theta, p, rho):
 def vasicek_lim_el(p, rho):
     """The expected loss of the large n limit of the Vasicek distribution.
 
-    :param p:
-    :param rho:
-    :return:
+    :param p: The probability of default
+    :param rho: The asset correlation (not needed)
+    :return: The expected default rate
     """
     return p
 
@@ -114,9 +125,9 @@ def vasicek_lim_el(p, rho):
 def vasicek_lim_ul(p, rho):
     """The unexpected loss of the large n limit of the Vasicek distribution.
 
-    :param p:
-    :param rho:
-    :return:
+    :param p:  The probability of default
+    :param rho: The asset correlation
+    :return: The default rate volatility
     """
     zmin = - settings.SCALE
     zmax = settings.SCALE
@@ -138,10 +149,10 @@ def vasicek_lim_ul(p, rho):
 def vasicek_lim_q(alpha, p, rho):
     """The quantile of the large-n Limit of the Vasicek distribution.
 
-    :param alpha:
-    :param p:
-    :param rho:
-    :return:
+    :param alpha: The desired quantile
+    :param p:   The probability of default
+    :param rho: The asset correlation
+    :return:  The default rate at that confidence level
     """
     a1 = stats.norm.ppf(p, loc=0.0, scale=1.0)
     a2 = stats.norm.ppf(alpha, loc=0.0, scale=1.0)